load place to edit && minor refactor
