+++ eyfo-backend/src/main/java/com/atstudio/eyfofalafel/backend/config/common/serialization/MappersConfig.java
@@ -1,20 +1,17 @@
-package com.atstudio.eyfofalafel.backend.config.common;
+package com.atstudio.eyfofalafel.backend.config.common.serialization;
 
-import com.atstudio.eyfofalafel.backend.service.location.LocationObjectFactory;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.SerializationFeature;
 import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
-import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+import org.zalando.logbook.HttpLogFormatter;
 
 @Configuration
-public class MappersConfig implements WebMvcConfigurer {
+public class MappersConfig {
 
     @Bean
-    @Qualifier("nullSkippingMapper")
     public ObjectMapper nullSkippingMapper() {
         ObjectMapper mapper = new ObjectMapper();
         mapper.registerModule(new JavaTimeModule());
@@ -24,8 +21,8 @@ public class MappersConfig implements WebMvcConfigurer {
     }
 
     @Bean
-    public LocationObjectFactory locationObjectFactory() {
-        return new LocationObjectFactory();
+    public HttpLogFormatter httpLogFormatter() {
+        return new PrincipalHttpLogFormatter(nullSkippingMapper());
     }
 
 }
